rename pkg/controllers/pluginconfig
