+++ pkg/controllers/plugin/helm_controller.go
@@ -1,7 +1,7 @@
 // SPDX-FileCopyrightText: 2024 SAP SE or an SAP affiliate company and Greenhouse contributors
 // SPDX-License-Identifier: Apache-2.0
 
-package pluginconfig
+package plugin
 
 import (
 	"context"
